+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to
 [#1417]: https://github.com/CosmWasm/cosmwasm/issues/1417
 [#1425]: https://github.com/CosmWasm/cosmwasm/pull/1425
 
+### Fixed
+
+- cosmwasm-vm: Bump `MODULE_SERIALIZATION_VERSION` to "v4" because the module
+  serialization format changed between Wasmer 2.2 and 2.3 ([#1426]).
+
+[#1426]: https://github.com/CosmWasm/cosmwasm/issues/1426
+
 ## [1.1.1] - 2022-09-15
 
 ### Fixed